change enrollment data type in table and model

+++ src/modules/users/infra/typeorm/entities/Student.ts
@@ -6,7 +6,7 @@ class Student{
   @PrimaryGeneratedColumn('uuid')
   id: string;
 
-  @Column('bigint')
+  @Column('int')
   enrollment_number: number;
 
   @Column('uuid')
